backend/requirements.txt
# Phase II Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.26
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Phase III Dependencies (Todo AI Chatbot)
mcp>=1.0.0
openai-agents-sdk>=0.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic-core>=2.0.0
websockets>=12.0
